🔵 Update from school : Adding more parsing functions

+++ src/window/window_size.c
@@ -6,7 +6,7 @@
 /*   By: jowagner <jowagner@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/05 18:21:55 by jowagner          #+#    #+#             */
-/*   Updated: 2025/03/06 15:27:29 by jowagner         ###   ########.fr       */
+/*   Updated: 2025/03/29 19:17:10 by jowagner         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 void	window_size(t_data *data)
 {
-	int i;
-	int j;
+	size_t	i;
+	size_t	j;
 
 	j = 0;
 	while (data->map[j] != NULL)
